rm file word_count.c
